tp JPA mysql h2

+++ src/main/java/com/bigcorp/booking/dao/FournisseurDao.java
@@ -17,7 +17,7 @@ public class FournisseurDao {
 		return merged;
 	}
 	
-	public Fournisseur findById(long id) {
+	public Fournisseur findById(Integer id) {
 		EntityManager em = PersistenceSingleton.INSTANCE.createEntityManager();
 		return em.find(Fournisseur.class, id);
 	}
